requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
pfzy==0.3.4
prompt-toolkit==3.0.36
Pygments==2.18.0
pytube==15.0.0
questionary==2.0.1
regex==2024.5.15
six==1.16.0
wcwidth==0.2.13
